migration: final cleanup and integration
