+++ README.md
@@ -1,68 +1,59 @@
+<!-- FlipDish Logo  -->
+<!-- <div align=center>
+    <img src="/src/assets/github/flipdish-logo-light.jpg" alt="FlipDish-Logo" title="DemoImage-home" width="250" height="150"> 
+</div> -->
+
+
 <!-- Introduction Text -->
 <div align="center">
-    <h1>Take-Home Assaignment</h1>
-    <h3>React app fetching data from API <br> endpoint & handeling JSON Data <h3>
+    <h2>Take-Home Assessment</h2>
+    <h4>NextJS app fetching data from API, utilizing Server-Side-Rendering <h4>
     <h3> 
       <a href='https://menu-technical-assignment.vercel.app/', target='_blank'>
-        <h5>live demo</h5>
+        <h5>live demo ↗</h5>
       <a/>
     </h3>
-        <h6>
-            built with <a href="https://vitejs.dev/" >Vite</a> &
-            hosted by <a href="https://vercel.com/">Vercel</a> 
-        </h6>
 </div>
 
----
+
 
 <h3 align='center'>
 Tech Used in this Project
 <h3>
 <p align='center'>
+        <img src="https://skillicons.dev/icons?i=next,ts,tailwind" /><br>
     <a href="https://skillicons.dev">
-        <img src="https://skillicons.dev/icons?i=vite,ts,tailwind" /><br>
         <img src="https://vitest.dev/logo.svg" width=50 alt="Next-Auth">
         <img src="https://skillicons.dev/icons?i=vercel,github" />
     </a>
 </p>
 
----
-
-<!-- DEMO IMAGE  -->
-<div align=center>
-    <img src="/src/assets/github/FlipDish-Demo-Desktop.png" alt="Demo-Desktop" title="DemoImage-home" width="600" height="400"> 
-    <img src="/src/assets//github//FlipDish-Demo-Mobile.png" alt="Demo-DemoMobile" title="DemoImage-login" width="300" height="500">    
-</div>
-<br>
-
-<!-- -------------------------------------------------------------------------- -->
-
-<h1 align='center'> Welcome & Introductory </h1>
-
-<!-- -------------------------------------------------------------------------- -->
-<hr/>
 
 ### Introduction:
 
 <hr/>
 <!-- -------------------------------------------------------------------------- -->
-Despite a hectic schedule, I took on the challenge of developing a React-based solution to render Flipdish restaurant menus. 
-This project was part of a time-constrained technical assessment, I spent approximetely between 5/6 hours in total on this assessment.
+This project is part of a time-constrained frontend technical assessment for FlipDish.  I dedicated approximately 5 to 6 hours to this assessment.
 
-Leveraging modern technologies including React, Typescript, and Tailwind CSS, I prioritized delivering essential functionality while ensuring clean, scalable code. You can explore my development journey through detailed comments on commits
+For this project, I prioritized leveraging Next.js for its server-side rendering capabilities, which are particularly beneficial for B2C companies.
 
-<h4><strong>⚠ Note for reviewer: </strong> </h4>
 
-You can find more information on my thought process - [here](https://github.com/DevonGifford/Menu--TechnicalAssignment#-development-thoughts-)
-<br>You can find open issues and future enhancements - [here](https://github.com/DevonGifford/Menu--TechnicalAssignment#-known-issues--potential-improvements--) <br/> Please feel free to ignore the code in the `src/components/ui` folder - this folder is auto generated by ShadCN and simply stores all the UI components from the library.
-<br/>
+> [!NOTE] 
+> **When reviewing;** <br/>
+> - Please feel free to ignore boilerplate ShadCN code  in the ' _components/ui_ '  <br/>
+> - You can find open/known issues & potential future enhancements [here](https://github.com/DevonGifford/Menu--TechnicalAssignment#-known-issues--potential-improvements--)
 
-<!-- -------------------------------------------------------------------------- -->
+
+<!-- DEMO IMAGE  -->
+<div align=center>
+    <img src="/src/assets/github/FlipDish-Demo-Mobile.png" alt="Demo-DemoMobile" title="DemoImage-login" width="150" height="250">    
+    <img src="/src/assets/github/FlipDish-Demo-Desktop.png" alt="Demo-Desktop" title="DemoImage-home" width="400" height="250"> 
+</div>
 <br>
-<hr>
 
-### 🔑 Key-Required Features of this project:
 
+
+### 🔑 Assessment Requirements:
 <hr>
 <!-- -------------------------------------------------------------------------- -->
 
@@ -76,7 +67,6 @@ You can find more information on my thought process - [here](https://github.com/
 
 ✅ Web app is fully Mobile/User-friendly and responsive.
 
-<br>
 
 #### Read and understand the structure of the Flipdish menu data object:
 
@@ -98,9 +88,10 @@ You can find more information on my thought process - [here](https://github.com/
 
 </br>
 </br>
-<!-- -------------------------------------------------------------------------- -->
-<hr>
-<h1 align='left'> Known issues & potential improvements: 🎯 </h1>
+
+
+
+<h3 align='left'>🎯 Known issues & potential improvements: </h3>
 <hr>
 
 <!-- -------------------------------------------------------------------------- -->
@@ -138,35 +129,9 @@ You can find more information on my thought process - [here](https://github.com/
 </details>
 <br><br>
 
-<!-- -------------------------------------------------------------------------- -->
 
-<hr>
-<h1 align='left'> Development Thoughts 🤔</h1>
-<hr>
 
-<!-- -------------------------------------------------------------------------- -->
-<details>
-<summary> Click here to expand </summary>
-<br/>
-
-#### [Commit #1](https://github.com/DevonGifford/Menu--TechnicalAssignment/commit/463c5d343ec247f0c052e9ef66fdc5b55cd65567#comments) - Thought Process on Tooling
-
-#### [Commit #2](https://github.com/DevonGifford/Menu--TechnicalAssignment/commit/89c78428dba057434fdfdf81951b22cff1e6bab2#comments) - Quick Prototype & Hosting
-
-#### [Commit #3](https://github.com/DevonGifford/Menu--TechnicalAssignment/commit/185e24303c7c7c81fe4c6913c503b0aebd348fd6#comments) - Cleaned up Prototype & hitting key requirement's
-
-#### [Commit #4](https://github.com/DevonGifford/Menu--TechnicalAssignment/commit/84c1cc639f1889943714c34e646bda475ffa074b#comments) - Testing completed and some final improvements
-
-#### [Commit #5](https://github.com/DevonGifford/Menu--TechnicalAssignment/commit/f1458cdfc1da1d62a51dc5852f1c43b06d38835d#comments) - Final Clean up
-
-## <br/>
-
-<!-- CLOSING DIV -->
-</details>
-<br><br>
-<!-- -------------------------------------------------------------------------- -->
-<hr>
-<h1 align='left'> Testing Results 🧪</h1>
+<h3 align='left'>🧪 Test Output: </h3>
 <hr>
 <!-- -------------------------------------------------------------------------- -->
 <details>
@@ -182,4 +147,4 @@ _Note: The code coverage shown above reflects the latest commit as as to provide
 
 <br><br>
 
-Spring Cleaning
+
