+++ src/app/layout.tsx
@@ -1,4 +1,9 @@
 import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+
+import "./globals.css";
+
+const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "FlipDish Test",
@@ -12,7 +17,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>
+      <body className={inter.className}>
         <div id="root">{children}</div>
       </body>
     </html>
