+++ src/components/ui/card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "./../../lib/utils";
+import { cn } from "../../utils/utils";
 
 /*
 🤔 Developer Note: mostly boilerplaye shadcn library - a few custom components
