+++ tailwind.config.js
@@ -2,10 +2,10 @@
 module.exports = {
   darkMode: ["class"],
   content: [
-    './pages/**/*.{ts,tsx}',
-    './components/**/*.{ts,tsx}',
-    './app/**/*.{ts,tsx}',
-    './src/**/*.{ts,tsx}',
+    './src/components/**/*.{ts,tsx}',
+    './src/utils/**/*.{ts,tsx}',
+    './src/types/**/*.{ts,tsx}',
+    './src/app/**/*.{ts,tsx}',
 	],
   theme: {
     container: {
