--- vite.config.ts
@@ -1,20 +1,0 @@
-/// <reference types="vitest" />
-
-import path from "path";
-import react from "@vitejs/plugin-react";
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  plugins: [react()],
-  test: { 
-    reporters: ['html', 'default'],
-    globals: true,
-    environment: 'happy-dom',
-    setupFiles: ['src/setupTest.ts']
-  },
-  resolve: {
-    alias: {
-      "@": path.resolve(__dirname, "./src"),
-    },
-  },
-});
